Moves business logic of MyAccountPage to MyAccountViewModel; Rename use***ViewModel to ***ViewModel
